requirements.txt
neat-python
pygame
numpy
matplotlib